+++ packages/escape-regex-str/.jsbits/example.js
@@ -2,10 +2,15 @@
 import $_NAME from '$_PACKAGE'
 
 //#set _TEMPL 'Unicorn (white): $'
-//#set _PARAM 'Unicorn (white): $1,000,000.00'
-//#set _RESULT RegExp(_TEMPL).test(_PARAM)
-const regexStr = $_NAME('$_TEMPL')
-const regex = new RegExp(regexStr)
+//#set _PARAM 'Unicorn (white): $1.00'
+const sourceStr = '$_TEMPL'
 
 // without $_NAME, the test fails.
-console.log(regex.test('$_PARAM')) // ⇒ $_RESULT
+//#set _RESULT RegExp(_TEMPL).test(_PARAM)
+const regex1 = new RegExp(sourceStr)
+console.log(regex1.test('$_PARAM')) // ⇒ $_RESULT
+
+// with $_NAME, it succeeds.
+//#set _RESULT RegExp(_F(_TEMPL)).test(_PARAM)
+const regex2 = new RegExp($_NAME(sourceStr))
+console.log(regex2.test('$_PARAM')) // ⇒ $_RESULT