+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog for @jsbits
 
+## [Unreleased]
+
+### Fixed
+
+- escape-regex-str: Bad written example showing incorrect result.
+
+
 ## [1.0.0] - 2018-11-09
 
 First public release, with 5 functions, CI tests and 100% coverage.